ai/requirements.txt
flask
flask-cors
gunicorn
psycopg2-binary
vaderSentiment
scikit-learn
numpy
pandas